requirements.txt
moviepy
opencv-python
path.py 